refactor: remove transcoding system and path mappings
Major architectural change: Peek now proxies all streams through Stash
instead of handling transcoding locally.

Server changes:
- Remove TranscodingManager.ts (~1200 lines) and ffmpegLogger.ts
- Simplify video.ts to only proxyStashStream and getCaption endpoints
- Remove PathMapping model from Prisma schema
- Simplify pathMapping.ts - keep only URL proxy functions
- Simplify setup controller/routes - only getSetupStatus and createFirstAdmin
- Remove transcoding stats from stats.ts
- Fix nodemon.json with explicit watch patterns (fixes ENOMEM warning)

Client changes:
- Refactor ServerSettings.jsx from ~1790 lines to ~117 lines
- Extract UserManagementSection, VersionInfoSection components
- Extract CreateUserModal, SyncFromStashModal components
- Remove Path Mappings section from ServerSettings
- Simplify SetupWizard to Welcome -> Create Admin -> Complete
- Remove path mapping API functions from setupApi
