+++ server/package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "serve": "node dist/index.js",
     "start": "node dist/index.js",
-    "dev": "nodemon --legacy-watch --watch './**/*.ts' --watch '../.env' --ext ts --exec 'node --loader ts-node/esm' index.ts",
+    "dev": "nodemon",
     "lint": "eslint .",
     "format": "prettier --write \"**/*.{js,ts,json,md}\"",
     "format:check": "prettier --check \"**/*.{js,ts,json,md}\"",
