+++ server/prisma/schema.prisma
@@ -129,16 +129,6 @@ model CustomTheme {
     @@index([userId])
 }
 
-model PathMapping {
-    id          Int         @id @default(autoincrement())
-    stashPath   String      // Path as reported by Stash (e.g., "/data", "C:\Videos")
-    peekPath    String      // Path where Peek accesses the files (e.g., "/app/media")
-    createdAt   DateTime    @default(now())
-    updatedAt   DateTime    @updatedAt
-
-    @@unique([stashPath])
-}
-
 model SceneRating {
     id          Int         @id @default(autoincrement())
     userId      Int
